+++ src/app.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ArtistModule } from './artist/artist.module';
-import { AlbumModule } from './album/album.module';
+import { ArtistModule } from './app/artist/artist.module';
+import { AlbumModule } from './app/album/album.module';
 import { DbModule } from './db/db.module';
-import { UserModule } from './user/user.module';
-import { TrackModule } from './track/track.module';
-import { FavoritesModule } from './favorites/favorites.module';
+import { UserModule } from './app/user/user.module';
+import { TrackModule } from './app/track/track.module';
+import { FavoritesModule } from './app/favorites/favorites.module';
 
 @Module({
   imports: [
