+++ src/app/album/interfaces/album.interface.ts
