+++ src/app/artist/interfaces/artist.interface.ts
