+++ src/app/favorites/favorites.module.ts
@@ -1,9 +1,9 @@
 import { Module, forwardRef } from '@nestjs/common';
 
-import { DbModule } from 'src/db/db.module';
-import { ArtistModule } from 'src/artist/artist.module';
-import { AlbumModule } from 'src/album/album.module';
-import { TrackModule } from 'src/track/track.module';
+import { DbModule } from '../../db/db.module';
+import { ArtistModule } from '../artist/artist.module';
+import { AlbumModule } from '../album/album.module';
+import { TrackModule } from '../track/track.module';
 import { FavoritesService } from './favorites.service';
 import { FavoritesController } from './favorites.controller';
