+++ src/app/favorites/favorites.service.ts
@@ -7,10 +7,10 @@ import {
 } from '@nestjs/common';
 import { validate as uuidValidate } from 'uuid';
 
-import { DbService } from 'src/db/db.service';
-import { ArtistService } from 'src/artist/artist.service';
-import { AlbumService } from 'src/album/album.service';
-import { TrackService } from 'src/track/track.service';
+import { DbService } from '../../db/db.service';
+import { ArtistService } from '../artist/artist.service';
+import { AlbumService } from '../album/album.service';
+import { TrackService } from '../track/track.service';
 
 @Injectable()
 export class FavoritesService {
