+++ src/app/favorites/interfaces/favorites.interface.ts
@@ -1,6 +1,6 @@
-import { Album } from 'src/album/interfaces/album.interface';
-import { Artist } from 'src/artist/interfaces/artist.interface';
-import { Track } from 'src/track/interfaces/track.interfaces';
+import { Album } from '../../album/interfaces/album.interface';
+import { Artist } from '../../artist/interfaces/artist.interface';
+import { Track } from '../../track/interfaces/track.interfaces';
 
 export interface Favorites {
   albums: string[]; // favorite albums ids
