+++ src/app/track/track.module.ts
@@ -1,7 +1,7 @@
 import { Module, forwardRef } from '@nestjs/common';
 
-import { DbModule } from 'src/db/db.module';
-import { FavoritesModule } from 'src/favorites/favorites.module';
+import { DbModule } from '../../db/db.module';
+import { FavoritesModule } from '../favorites/favorites.module';
 import { TrackService } from './track.service';
 import { TrackController } from './track.controller';
