+++ src/app/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { DbModule } from 'src/db/db.module';
+import { DbModule } from '../../db/db.module';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
