+++ src/db/db.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
 
-import { Album } from 'src/album/interfaces/album.interface';
-import { Artist } from 'src/artist/interfaces/artist.interface';
-import { Favorites } from 'src/favorites/interfaces/favorites.interface';
-import { Track } from 'src/track/interfaces/track.interfaces';
-import { User } from 'src/user/interfaces/user.interface';
+import { Album } from 'src/app/album/interfaces/album.interface';
+import { Artist } from 'src/app/artist/interfaces/artist.interface';
+import { Favorites } from 'src/app/favorites/interfaces/favorites.interface';
+import { Track } from 'src/app/track/interfaces/track.interfaces';
+import { User } from 'src/app/user/interfaces/user.interface';
 
 @Injectable()
 export class DbService {
